pyproject.toml
[project]
name = "landing-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "logfire[fastapi]>=3.14.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.5",
    "uvicorn>=0.34.0",
]
